requirements.txt
altair==5.3.0
matplotlib==3.9.1
numpy==2.0.0
opencv_python==4.10.0.84
pandas==2.2.2
Pillow==10.4.0
scikit_learn==1.5.1
seaborn==0.13.2
Shapely==2.0.4
streamlit==1.36.0
torch==2.3.1
ultralytics==8.2.51
